State a conversion between what is rendered and what the boundary needs, never against the machine width

+++ tests/corpus/raw-baseline/arm64_O0/murmur3_32.c
@@ -20,8 +20,8 @@ uint32_t sym__murmur3_32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
         for (stack_m64 = 0; stack_m64 < X8_4; stack_m64++) {
             uint32_t tmp_25180_2 = ((uint32_t*)stack_m24)[stack_m64];
             uint32_t X0_3 = sym__rotl32((uint32_t)(tmp_25180_2 * 0xcc9e2d51), 15);
-            uint32_t X0_5 = sym__rotl32((uint32_t)(tmp_24d00_1 ^ (uint32_t)X0_3 * 0x1b873593), 13);
-            tmp_24d00_1 = (uint32_t)X0_5 * 5 - 0x19ab949c;
+            uint32_t X0_5 = sym__rotl32((uint32_t)(tmp_24d00_1 ^ X0_3 * 0x1b873593), 13);
+            tmp_24d00_1 = X0_5 * 5 - 0x19ab949c;
         }
         {
             uint32_t stack_m84;
@@ -46,7 +46,7 @@ uint32_t sym__murmur3_32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
             {
                 uint8_t tmp_25500_5 = *tmp_12480_1;
                 uint32_t X0_7 = sym__rotl32((uint32_t)(((uint32_t)tmp_25500_5 ^ stack_m84) * 0xcc9e2d51), 15);
-                tmp_24d00_1 ^= (uint32_t)X0_7 * 0x1b873593;
+                tmp_24d00_1 ^= X0_7 * 0x1b873593;
             }
             {
 L3: ;
