+++ tests/corpus/raw-baseline/arm64_O0/xxhash32.c
@@ -38,25 +38,25 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                     X8_5 = stack_m24;
                     uint32_t tmp_24c00_5 = *(uint32_t*)X8_5;
                     uint32_t X0_3 = sym__rotl32((uint32_t)(stack_m68 - tmp_24c00_5 * 0x7a143589), 13);
-                    stack_m68 = (uint32_t)X0_3 * 0x9e3779b1;
+                    stack_m68 = X0_3 * 0x9e3779b1;
                     X8_5 = stack_m24;
                     stack_m24 = X8_5 + 4;
                     X8_5 = stack_m24;
                     uint32_t tmp_24c00_11 = *(uint32_t*)X8_5;
                     uint32_t X0_5 = sym__rotl32((uint32_t)(stack_m72 - tmp_24c00_11 * 0x7a143589), 13);
-                    stack_m72 = (uint32_t)X0_5 * 0x9e3779b1;
+                    stack_m72 = X0_5 * 0x9e3779b1;
                     X8_5 = stack_m24;
                     stack_m24 = X8_5 + 4;
                     X8_5 = stack_m24;
                     uint32_t tmp_24c00_17 = *(uint32_t*)X8_5;
                     uint32_t X0_7 = sym__rotl32((uint32_t)(tmp_24d00_1 - tmp_24c00_17 * 0x7a143589), 13);
-                    tmp_24d00_1 = (uint32_t)X0_7 * 0x9e3779b1;
+                    tmp_24d00_1 = X0_7 * 0x9e3779b1;
                     X8_5 = stack_m24;
                     stack_m24 = X8_5 + 4;
                     X8_5 = stack_m24;
                     uint32_t tmp_24c00_23 = *(uint32_t*)X8_5;
                     uint32_t X0_9 = sym__rotl32((uint32_t)(stack_m80 - tmp_24c00_23 * 0x7a143589), 13);
-                    stack_m80 = (uint32_t)X0_9 * 0x9e3779b1;
+                    stack_m80 = X0_9 * 0x9e3779b1;
                     X8_5 = stack_m24;
                     stack_m24 = X8_5 + 4;
                     {
@@ -72,7 +72,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                     uint32_t X0_13 = sym__rotl32((uint32_t)stack_m72, 7);
                     uint32_t X0_16 = sym__rotl32((uint32_t)tmp_24d00_1, 12);
                     uint32_t X0_18 = sym__rotl32((uint32_t)stack_m80, 18);
-                    stack_m52 = (uint32_t)X0_11 + (uint32_t)X0_13 + (uint32_t)X0_16 + (uint32_t)X0_18;
+                    stack_m52 = X0_11 + X0_13 + X0_16 + X0_18;
                 }
             }
         }
@@ -90,7 +90,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                     X8_5 = stack_m24;
                     uint32_t tmp_24c00_36 = *(uint32_t*)X8_5;
                     uint32_t X0_22 = sym__rotl32((uint32_t)(stack_m52_2 - tmp_24c00_36 * 0x3d4d51c3), 17);
-                    stack_m52_2 = (uint32_t)X0_22 * 0x27d4eb2f;
+                    stack_m52_2 = X0_22 * 0x27d4eb2f;
                     X8_5 = stack_m24;
                     stack_m24 = X8_5 + 4;
                 }
@@ -104,7 +104,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                         X8_5 = stack_m24;
                         uint8_t tmp_25500_2 = *(uint8_t*)X8_5;
                         uint32_t X0_25 = sym__rotl32((uint32_t)(stack_m52_2 + (uint32_t)tmp_25500_2 * 0x165667b1), 11);
-                        stack_m52_2 = (uint32_t)X0_25 * 0x9e3779b1;
+                        stack_m52_2 = X0_25 * 0x9e3779b1;
                         X8_5 = stack_m24;
                         stack_m24 = X8_5 + 1;
                     }
