+++ tests/corpus/raw-baseline/arm64_O1/crc32_bitwise.c
@@ -12,10 +12,10 @@ uint64_t sym__crc32_bitwise(uint64_t X0_0, uint64_t X1_0)
         for (; ; ) {
             uint32_t X11_5;
             uint8_t tmp_25600_2 = ((uint8_t*)X0_0)[X9_1];
-            X8_1 = (uint32_t)((uint32_t)X8_1 ^ (uint32_t)tmp_25600_2);
+            X8_1 = (uint32_t)(X8_1 ^ (uint32_t)tmp_25600_2);
             X11_5 = 8;
             for (; ; ) {
-                tmp_20380_4 = (uint32_t)X8_1 >> 1 ^ (((uint32_t)X8_1 & 1) * 0xffffffff & 0xedb88320);
+                tmp_20380_4 = X8_1 >> 1 ^ ((X8_1 & 1) * 0xffffffff & 0xedb88320);
                 X8_1 = (uint32_t)tmp_20380_4;
                 uint32_t tmp_lane_1000006d0_a_8_1 = X11_5;
                 X11_5 = (uint32_t)(tmp_lane_1000006d0_a_8_1 - 1);
