+++ tests/corpus/raw-baseline/x64_O1/adler32.c
@@ -14,7 +14,7 @@ uint64_t sym__adler32(uint64_t RDI_0, uint64_t RSI_0)
         RAX_1 = (uint64_t)0;
         for (; ; ) {
             uint8_t tmp_11e00_2 = ((uint8_t*)RDI_0)[RCX_1];
-            uint32_t tmp_lane_100000650_7_9_1 = (uint32_t)tmp_11e00_2 + (uint32_t)R8_1;
+            uint32_t tmp_lane_100000650_7_9_1 = (uint32_t)tmp_11e00_2 + R8_1;
             tmp_lane_100000650_49_11_1 = tmp_lane_100000650_7_9_1 - (uint32_t)((uint64_t)(int32_t)((uint64_t)tmp_lane_100000650_7_9_1 * 0x80078071 >> 47) * 0xfff1);
             R8_1 = (uint32_t)tmp_lane_100000650_49_11_1;
             uint32_t tmp_lane_100000650_53_15_1 = (uint32_t)RAX_1 + tmp_lane_100000650_49_11_1;
